Cache component lookups
