+++ compiler/src/main/java/com/squareup/anvil/compiler/ClassScannerKsp.kt
@@ -1,24 +1,33 @@
 package com.squareup.anvil.compiler
 
 import com.google.devtools.ksp.KspExperimental
+import com.google.devtools.ksp.getVisibility
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.symbol.KSPropertyDeclaration
 import com.google.devtools.ksp.symbol.KSType
 import com.google.devtools.ksp.symbol.KSTypeReference
 import com.google.devtools.ksp.symbol.Origin
+import com.google.devtools.ksp.symbol.Visibility
 import com.squareup.anvil.compiler.ClassScannerKsp.GeneratedProperty.ReferenceProperty
 import com.squareup.anvil.compiler.ClassScannerKsp.GeneratedProperty.ScopeProperty
 import com.squareup.anvil.compiler.api.AnvilCompilationException
+import com.squareup.anvil.compiler.codegen.ksp.KspAnvilException
 import com.squareup.anvil.compiler.codegen.ksp.fqName
+import com.squareup.anvil.compiler.codegen.ksp.isInterface
 import com.squareup.anvil.compiler.codegen.ksp.resolveKSClassDeclaration
+import com.squareup.anvil.compiler.codegen.ksp.returnTypeOrNull
 import com.squareup.anvil.compiler.codegen.ksp.scope
 import com.squareup.kotlinpoet.ksp.toClassName
 import org.jetbrains.kotlin.name.FqName
 
 internal class ClassScannerKsp {
 
-  private val cache = mutableMapOf<CacheKey, Collection<List<GeneratedProperty>>>()
+  private val generatedPropertyCache = mutableMapOf<CacheKey, Collection<List<GeneratedProperty>>>()
+  private val parentComponentCache = mutableMapOf<FqName, KSClassDeclaration?>()
+  private val overridableParentComponentFunctionCache =
+    mutableMapOf<FqName, List<KSFunctionDeclaration>>()
 
   /**
    * Externally-contributed contributions, which are important to track so that we don't try to
@@ -46,7 +55,7 @@ internal class ClassScannerKsp {
     scope: KSType?,
   ): Sequence<KSClassDeclaration> {
     val propertyGroups: Collection<List<GeneratedProperty>> =
-      cache.getOrPut(CacheKey(annotation, resolver.hashCode())) {
+      generatedPropertyCache.getOrPut(CacheKey(annotation, resolver.hashCode())) {
         resolver.getDeclarationsFromPackage(HINT_PACKAGE)
           .filterIsInstance<KSPropertyDeclaration>()
           .mapNotNull { GeneratedProperty.from(it) }
@@ -89,7 +98,8 @@ internal class ClassScannerKsp {
         // Check that the annotation really is present. It should always be the case, but it's
         // a safetynet in case the generated properties are out of sync.
         clazz.annotations.any {
-          it.annotationType.resolve().toClassName().fqName == annotation && (scope == null || it.scope() == scope)
+          it.annotationType.resolve()
+            .toClassName().fqName == annotation && (scope == null || it.scope() == scope)
         }
       }
       .onEach { clazz ->
@@ -142,4 +152,91 @@ internal class ClassScannerKsp {
     val fqName: FqName,
     val resolverHash: Int,
   )
+
+  /**
+   * Finds the applicable parent component interface (if any) contributed to this component.
+   */
+  fun findParentComponentInterface(
+    componentClass: KSClassDeclaration,
+    factoryClass: KSClassDeclaration?,
+    parentScopeType: KSType?,
+  ): KSClassDeclaration? {
+    val fqName = componentClass.fqName
+
+    // Can't use getOrPut because it doesn't differentiate between absent and null
+    if (fqName in parentComponentCache) {
+      return parentComponentCache[fqName]
+    }
+
+    val contributedInnerComponentInterfaces = componentClass
+      .declarations
+      .filterIsInstance<KSClassDeclaration>()
+      .filter(KSClassDeclaration::isInterface)
+      .filter { nestedClass ->
+        nestedClass.annotations
+          .any {
+            it.fqName == contributesToFqName && (if (parentScopeType != null) it.scope() == parentScopeType else true)
+          }
+      }
+      .toList()
+
+    val componentInterface = when (contributedInnerComponentInterfaces.size) {
+      0 -> return null
+      1 -> contributedInnerComponentInterfaces[0]
+      else -> throw KspAnvilException(
+        node = componentClass,
+        message = "Expected zero or one parent component interface within " +
+          "${componentClass.fqName} being contributed to the parent scope.",
+      )
+    }
+
+    val functions = overridableParentComponentFunctions(
+      componentInterface,
+      componentClass.fqName,
+      factoryClass?.fqName,
+    )
+
+    when (functions.count()) {
+      0 -> return null
+      1 -> {
+        // This is ok
+      }
+      else -> throw KspAnvilException(
+        node = componentClass,
+        message = "Expected zero or one function returning the " +
+          "subcomponent ${componentClass.fqName}.",
+      )
+    }
+
+    parentComponentCache[fqName] = componentInterface
+    return componentInterface
+  }
+
+  /**
+   * Returns a list of overridable parent component functions from a given [parentComponent]
+   * for the given [targetReturnType].
+   */
+  fun overridableParentComponentFunctions(
+    parentComponent: KSClassDeclaration,
+    targetReturnType: FqName,
+    factoryClass: FqName?,
+  ): List<KSFunctionDeclaration> {
+    val fqName = parentComponent.fqName
+
+    // Can't use getOrPut because it doesn't differentiate between absent and null
+    if (fqName in overridableParentComponentFunctionCache) {
+      return overridableParentComponentFunctionCache.getValue(fqName)
+    }
+
+    return parentComponent.getAllFunctions()
+      .filter { it.isAbstract && it.getVisibility() == Visibility.PUBLIC }
+      .filter {
+        val returnType = it.returnTypeOrNull()?.resolveKSClassDeclaration() ?: return@filter false
+        returnType.fqName == targetReturnType || (factoryClass != null && returnType.fqName == factoryClass)
+      }
+      .toList()
+      .also {
+        overridableParentComponentFunctionCache[fqName] = it
+      }
+  }
 }
